Makes PayPalButton safe for SSR
If the component was rendered on the server, the old code would eval `!isSdkReady`, find it to be true and then because of the `&&` it would also have to check if `window.paypal` were true. That would fail because `window` would be undefined on the server. Switching to an `||` means that as soon as `!isSdkReady` evaluates to true, it doesn't need to evaluation the rest of the statement and `null` is returned.